# Optimize Half.Equals