--- a/src/libraries/System.Private.CoreLib/src/System/Half.cs
+++ b/src/libraries/System.Private.CoreLib/src/System/Half.cs
@@ -424,12 +424,8 @@ namespace System
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         public bool Equals(Half other)
         {
-            if (_value == other._value)
-            {
-                return true;
-            }
-
-            return AreZero(this, other)
+            return _value == other._value
+                || AreZero(this, other)
                 || (IsNaN(this) && IsNaN(other));
         }